A noun is a Person, Place, Thing, or Idea.
sprites stay in a designated location on the map. A multithread called
Wait now controls walking speed. Battles can kill you- try not to die as
the results are not friendly. walking on city tiles pays 20 g if you
have it and are damaged then heals you. monsters are more formidable,
you no longer can defend to invincibility. battles are now turn based.
spacebar is the new x button in gaming, use it for everything. player
menu by pressing enter is more refined.