SPRITES NOW MOVE CORRECTLY RELATIVE TO YOU!HWVER...
HOWEVER...(HAD TO MISSPELL IT BECAUSE THE CHAR LIMIT)...When retrieving
x coordinate of sprite, it is correct (so if I get player x and sprite x
and I stand on top of the sprite, they are the same!), but when doing
the same with y, the sprite's y is lower than player y by 15! (15 is
half of 30, which is starty) can't figure out why this is.