Cleanup and alignment of monster name
hud.drawAlignedString(g, m.getName(), 800/2, 30+100);

hud.drawAlignedString
will draw a specified string at the x, y coordinate you specify. it
automatically aligns it by using fontmetrics to figure out the width of
the string and subtracts off half of that from the x coordinate, so the
center of the string is put at the (x,y) you specify! very useful