Graphics and functionality changes
- see all my commits for today (since your "some spelling corrections"
commit) for full list
- hud has tons of drawCentered___ methods which will draw centered crap,
so you don't have to worry about the size of, say, an image affecting
the alignment on the screen
- put tabs in some places to make it look more prettyfuler
- spaced out attack, spell, item, defend, run buttons and used
drawCentered methods to make it easier to see what's being drawn
where...however, you'll see a "-22" in some of them. somehow, they're
being shifted by 22, so I need to unshift by that much.
- when health, mana, etc low, it changes to a dark color so you can see
it against white background
- resized bars by 2 pixels to make it look better (subtle)
- beds now restore mana (I remember you mentioning this as a "to do")
- setMana function in player