Small changes - read desc
updated npc info when you load the game.
hud does this automatically when you enter a town. so, for example, if
you save in a town, then exit and open the game again, load, you'll be
in the town. that works just fine. but if you try to talk to an npc,
you'll get a nullptrexception since npcdata wasn't loaded because you
didn't step on the tile for towns. this is fixed now.

commented out save/load from within battle spell/item menus