Each tile now has 3 "values"
first value is normal tile number using getVal(), second two are special
ones that we talked about that one day (making teleportation easier and
using special tile attributes).

Okay, so breakdown:

Original tile map looked like "A A A B A A A..."

If a tile didn't exist in the text file, the matrix element would be set
to -1, so we know which ones are out of bounds (if it gets past the
error checking). Same thing happens with the 2 extra special values, so
you don't even have to add them if you're not going to use them.

Suppose you add some, then you'd see "AAA AAA ABA ABB..." or something
like that, where the last 2 letters of each tile are the special values.
Or it could look like "A A ABA ABB" if the first two tiles were normal
and you didn't want to do anything special with them.

The corresponding python files will automatically put an extra AA at the
end of all your tiles, but from previous paragraph, this is not
necessary, so I can take it out later.

You can get these extra values by using the methods in the TileMap
class, just like you'd use the getVal(x,y) method in that class. Feel
free to rename the other two arrays (right now they're just "special1"
and "special2", not very descriptive).

I believe that once we're able to implement the extra values into the
rest of the game, town traveling will be much easier, and we won't have
to hardcode the maptracker to say that if you're in map x, then go to
map y and vice versa. A few commits ago, I moved the maptracker inside
of the player, so the player now holds the info for what map he's in. I
believe the methods are getMapTracker() and setMapTracker(), although
you can always just look it up.

Oh, one last thing I did to make reading the maps easier. Instead of
reading in the maps in a customized way, I offloaded it to the Load
class. Since the maps were delimited by spaces instead of tabs, I made
the readFileToArray() method in load require 2 parameters instead of 1
-- fileName AND delimiter. For almost everything, delimiter is "\t", but
for map, I used " ". I guess we can move everything to use tabs instead,
but I didn't want to fiddle with the map files too much.