adding back 3rd special
each map will hold 3 values now, the last one will be the spawn rate for
that location, in case we want a zubat-like cave or something, as well
as not having a ton of dont battle conditions. Just as well this will
allow if we want teleporters or something, so it will be 1-teleport
command:: 2-current tile:: 3-destination tile.