Created effects class
effects class is modular, so you can add effects in. right now I have a
rumble-like effect (earthquake). I can add some that tint the color of
what's being drawn. all you need to do is initialize it somewhere (for
this example, I initialized it with the a strength of 10, and a duration
of 1000 milliseconds. When the map is drawn, it will use the default
effect (earthquake) to draw the map until the duration passes, after
which it will just draw it normally.

Soon, the initEffectParams function will take a string for which effect
to use ("earthquake", "none", "darkness", "redtint", or stuff like
that), and the corresponding effect will be used through a switch
statement inside of draw()

Currently, the ground shakes when you walk (like a boss). We could
easily make that happen when you hit an edge, or only for certain tiles,
etc. Ze pozibility ze are endless