new effect
gradient. I think this would be nice (or something close to it) for the
teleportation. we can make a player spinning sprite or something, and
finally, we need to delay the teleportation. I thought about putting a
thread delay, but since everything is on the same thread I think, then
if we put a delay, music, animations, etc will all pause