changes to effect
draw method much smaller (one for buffered image, one without), so I
moved all the effect handling into handleEffects()

player now has isBlocked() function that will return true if the
direction he's going is blocked,--it's like canMoveLeft, canMoveRight,
etc all combined so it's in general if he can't move in the direction
he's facing.

I used that to make the switch case in RPG much shorter, and to add a
shake effect when you go to a tile you can't walk on. Gradient now
appears when you change maps or go on an umbrella. feel free to change
that, was just playing around with it.

eff.setEffect(""); or eff.setEffect("none"); clears all effects, so the
map gets drawn normally or whatever (if not a map)