map changes
the number of sprites displayed on a map is determined by the 4th and
5th special values, I am thinking the background music for each city can
be displayed by the 6th value. for some reason the sprites in the second
city are not displaying messaging properly. I am confused. the second
city is to the far south, you can take the bottom left most teleporter
to get to it.