maps made easy. then really hard.
i forgot what all I did, but basically now when you enter a map it will
look for the corresponding exit to go to the same place you started and
make your starting coords that. dying will respawn you to the last bed
you slept in. and battle frequency is controlled in player. there will
be new commands soon- a tile can contain a chest with gold or equipment,
and can initiate a boss fight. new tiles for making dungeons. new maps.
the bottom most right city (in the worst area) has a dungeon in the
city. music for every occasion also soon to follow. hope to wrap it all
up before 1:30